chore: moved operational solution subtask status enum into operational_solution_subtask.go
